feat: Code compiling

+++ CUDA/includes/ray.cuh
@@ -21,7 +21,7 @@ class ray {
         __device__ point3 origin() const  { return orig; }
         __device__ vec3 direction() const { return dir; }
 
-        __device__ point3 at(double t) const {
+        __device__ point3 at(float t) const {
             return orig + t*dir;
         }
 };
