+++ CUDA/src/sphere.cu
@@ -1,26 +1,26 @@
 #include "shape.cuh"
 
-__host__ Sphere::Sphere(point3 center, double rad, Material* m){
+__host__ Sphere::Sphere(point3 center, float rad, Material* m){
     this->position = center;
     this->radius = rad;
     this->obj_material = m;
 }
 
-__device__ double Sphere::hit(const ray& r, float t_min, float t_max){
-    double t = 0;
+__device__ float Sphere::hit(const ray& r, float t_min, float t_max){
+    float t = 0;
     vec3 oc = r.origin() - this->position;
     auto a = dot(r.direction(), r.direction());
-    auto b = 2.0 * dot(oc, r.direction());
+    auto b = 2.0f * dot(oc, r.direction());
     auto c = dot(oc, oc) - radius*radius;
     auto discriminant = b*b - 4*a*c;
 
     if(discriminant < 0)
-        return -1.0;
+        return -1.0f;
     
     t = (-b-sqrt(discriminant))/(2.0*a);
 
     if(t > t_max || t < t_min)
-        return -1.0;
+        return -1.0f;
 
     return t;
 }
